+++ lib/macro.dart
@@ -4,6 +4,7 @@ export 'package:collection/collection.dart' show DeepCollectionEquality;
 
 export 'src/analyzer/error.dart';
 export 'src/client/run.dart';
+export 'src/common/models.dart' show AutoRebuildResult;
 export 'src/core/core.dart' hide MacroX, BaseMacroGenerator;
 export 'src/core/extension.dart';
 export 'src/core/modifier.dart';